# public/openapi-schema.yaml
# Généré par `pnpm cli openapi:generate` — ne pas éditer à la main.
components:
  schemas:
    Demande:
      type: object
      properties:
        batiment:
          type: object
          properties:
            energie_chauffage:
              anyOf:
                - type: string
                  enum:
                    - "Électricité"
                    - Gaz
                    - Fioul
                    - Autre / Je ne sais pas
                - type: "null"
              description: "Énergie de chauffage actuelle : Électricité, Gaz, Fioul, Autre / Je ne sais pas."
            etablissement:
              anyOf:
                - type: string
                - type: "null"
              description: "Nom de l'établissement ou de la structure."
            nombre_logements:
              anyOf:
                - type: integer
                - type: "null"
              description: "Nombre de logements (valeur corrigée par le gestionnaire si renseignée)."
            surface_m2:
              anyOf:
                - type: number
                - type: "null"
              description: "Surface chauffée (m²)."
            type_chauffage:
              anyOf:
                - type: string
                  enum:
                    - Individuel
                    - Collectif
                    - Autre / Je ne sais pas
                - type: "null"
              description: "Type de chauffage : Individuel, Collectif, Autre / Je ne sais pas."
            type_structure:
              anyOf:
                - type: string
                  enum:
                    - Tertiaire
                    - "Copropriété"
                    - Bailleur social
                    - Maison individuelle
                    - Autre
                - type: "null"
              description: "Type de structure : Tertiaire, Copropriété, Bailleur social, Maison individuelle, Autre."
          required:
            - energie_chauffage
            - etablissement
            - nombre_logements
            - surface_m2
            - type_chauffage
            - type_structure
          additionalProperties: false
          description: "Caractéristiques du bâtiment."
        commentaire:
          anyOf:
            - type: string
            - type: "null"
          description: Commentaire de suivi du gestionnaire. Modifiable.
        contact:
          type: object
          properties:
            email:
              anyOf:
                - type: string
                - type: "null"
              description: E-mail du prospect.
            nom:
              anyOf:
                - type: string
                - type: "null"
              description: Nom du prospect.
            prenom:
              anyOf:
                - type: string
                - type: "null"
              description: "Prénom du prospect."
            telephone:
              anyOf:
                - type: string
                - type: "null"
              description: "Téléphone du prospect."
          required:
            - email
            - nom
            - prenom
            - telephone
          additionalProperties: false
          description: "Coordonnées du prospect — données personnelles (RGPD)."
        date_creation:
          type: string
          description: "Date d'enregistrement de la demande (ISO 8601)."
        date_modification:
          type: string
          description: "Date de dernière modification (ISO 8601). Sert de curseur de polling."
        eligibilite:
          type: object
          properties:
            dans_pdp:
              type: boolean
              description: "Située dans un Périmètre de Développement Prioritaire."
            distance_reseau_m:
              anyOf:
                - type: number
                - type: "null"
              description: "Distance estimée au réseau (m)."
          required:
            - dans_pdp
            - distance_reseau_m
          additionalProperties: false
          description: "Éligibilité au moment de la demande."
        id:
          type: string
          description: "Identifiant unique et stable de la demande côté FCU."
        localisation:
          type: object
          properties:
            adresse:
              anyOf:
                - type: string
                - type: "null"
              description: Adresse du bien.
            code_postal:
              anyOf:
                - type: string
                - type: "null"
              description: Code postal.
            commune_code:
              anyOf:
                - type: string
                - type: "null"
              description: Code INSEE de la commune.
            commune_label:
              anyOf:
                - type: string
                - type: "null"
              description: Nom de la commune.
            departement_code:
              anyOf:
                - type: string
                - type: "null"
              description: "Code INSEE du département."
            departement_label:
              anyOf:
                - type: string
                - type: "null"
              description: "Nom du département."
            latitude:
              anyOf:
                - type: number
                - type: "null"
              description: Latitude (WGS84).
            longitude:
              anyOf:
                - type: number
                - type: "null"
              description: Longitude (WGS84).
            region_code:
              anyOf:
                - type: string
                - type: "null"
              description: "Code INSEE de la région."
            region_label:
              anyOf:
                - type: string
                - type: "null"
              description: "Nom de la région."
          required:
            - adresse
            - code_postal
            - commune_code
            - commune_label
            - departement_code
            - departement_label
            - latitude
            - longitude
            - region_code
            - region_label
          additionalProperties: false
          description: "Localisation du bien — donnée personnelle (RGPD)."
        reseau:
          type: object
          properties:
            gestionnaire:
              anyOf:
                - type: string
                - type: "null"
              description: "Gestionnaire du réseau affecté."
            id_fcu:
              type: integer
              description: "Identifiant interne FCU du réseau affecté."
            identifiant_sncu:
              anyOf:
                - type: string
                - type: "null"
              description: "Identifiant national SNCU (null pour un réseau en construction)."
            maitre_ouvrage:
              anyOf:
                - type: string
                - type: "null"
              description: "Maître d'ouvrage du réseau (réseaux de chaleur uniquement, sinon null)."
            nom:
              anyOf:
                - type: string
                - type: "null"
              description: "Nom du réseau affecté."
            type:
              type: string
              enum:
                - reseau_de_chaleur
                - reseau_en_construction
              description: reseau_de_chaleur ou reseau_en_construction.
          required:
            - gestionnaire
            - id_fcu
            - identifiant_sncu
            - maitre_ouvrage
            - nom
            - type
          additionalProperties: false
          description: "Réseau affecté à la demande."
        statut:
          type: string
          enum:
            - "À traiter"
            - "Non réalisable"
            - "Recontacté pour étude"
            - "Proposition commerciale envoyée"
            - "Voté en AG"
            - Travaux en cours
            - "Réalisé"
            - "Projet abandonné par le prospect"
          description: "Statut d'avancement : À traiter, Non réalisable, Recontacté pour étude, Proposition commerciale envoyée, Voté en AG, Travaux en cours, Réalisé, Projet abandonné par le prospect."
      required:
        - batiment
        - commentaire
        - contact
        - date_creation
        - date_modification
        - eligibilite
        - id
        - localisation
        - reseau
        - statut
      additionalProperties: false
      description: "Demande de raccordement (contrat partenaire stable, versionné, indépendant du stockage interne)."
  securitySchemes:
    bearerAuth:
      description: "Token d'organisation fourni par l'équipe FCU (en-tête Authorization: Bearer fcu_…)."
      scheme: bearer
      type: http
info:
  contact:
    name: Equipe France Chaleur Urbaine
    url: "https://france-chaleur-urbaine.beta.gouv.fr/contact"
  description: "L'API France Chaleur Urbaine permet de télécharger les données et tracés des réseaux de chaleur, tester la proximité avec les réseaux, et (API partenaire authentifiée) synchroniser les demandes de raccordement avec un CRM."
  license:
    name: Licence Ouverte / Open Licence version 2.0
    url: "https://www.etalab.gouv.fr/licence-ouverte-open-licence/"
  termsOfService: "https://france-chaleur-urbaine.beta.gouv.fr/mentions-legales"
  title: API France Chaleur Urbaine
  version: "1.0.0"
openapi: "3.1.0"
paths:
  "/v1/eligibility":
    get:
      parameters:
        - description: "Latitude (EPSG:4326)"
          in: query
          name: lat
          required: true
          schema:
            type: number
        - description: "Longitude (EPSG:4326)"
          in: query
          name: lon
          required: true
          schema:
            type: number
      responses:
        "200":
          content:
            application/json:
              schema:
                properties:
                  distance:
                    description: "distance à vol d'oiseau entre l'adresse testée et le réseau de chaleur le plus proche"
                    type: number
                  futurNetwork:
                    description: "true si le réseau le plus proche, sur lequel est établi le résultat isEligible, est un réseau actuellement en construction"
                    type: boolean
                  gestionnaire:
                    description: "nom du gestionnaire du réseau le plus proche"
                    type: string
                  id:
                    description: "identifiant national du réseau de chaleur le plus proche"
                    type: string
                  inPDP:
                    description: "true si l'adresse testée se situe dans le périmètre de développement prioritaire d'un réseau classé (zone où s'applique une obligation de raccordement pour certains bâtiments)"
                    type: boolean
                  isEligible:
                    description: "true si la distance entre l'adresse testée et le réseau de chaleur le plus proche est inférieure à un certain seuil de distance* ou si nous avons connaissance d'un réseau de chaleur dans le quartier dont nous n'avons pas le tracé. * Le seuil de distance est fixé à 100 m sur Paris, 200 m ailleurs. En pratique ce seuil dépendra des besoins en chaleur du bâtiment à raccorder. Les valeurs prises ici sont volontairement élevées : le gestionnaire du réseau sera le seul à même d'évaluer la faisabilité du raccordement"
                    type: boolean
                  name:
                    example: Paris et communes limitrophes
                    type: string
                  rateCO2:
                    description: "contenu CO2 en analyse du cycle de vie (émissions directes et indirectes) du réseau de chaleur le plus proche, en kg/kWh, sur la base du dernier arrêté \"DPE\""
                    type: number
                  rateENRR:
                    description: "taux d'énergies renouvelables et de récupération du réseau de chaleur le plus proche, en %, sur la base du dernier arrêté \"DPE\""
                    type: number
                type: object
          description: "Opération réussie"
        "400":
          description: "Erreur de paramètres"
        "500":
          description: Erreur technique
      summary: "Tester l'éligibilité d'un point géographique avec les réseaux de chaleur."
      tags:
        - "Informations générales"
  "/v1/networks":
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  properties:
                    contenu CO2:
                      example: 0.154
                      type: number
                    contenu CO2 ACV:
                      example: 0.188
                      type: number
                    Gestionnaire:
                      example: CPCU (ENGIE SOLUTIONS)
                      type: string
                    geom:
                      description: "Géométrie au format GeoJSON"
                      properties:
                        coordinates:
                          items:
                            items:
                              items:
                                type: number
                              type: array
                            type: array
                          type: array
                        type:
                          example: MultiLineString
                          title: "Type de géométrie"
                          type: string
                      type: object
                    Identifiant reseau:
                      description: identifiant SNCU
                      example: "7501C"
                      type: string
                    id_fcu:
                      description: "identifiant interne à France Chaleur Urbaine"
                      type: number
                    nom_reseau:
                      example: Paris et communes limitrophes
                      type: string
                    "Taux EnR&R":
                      example: 50.1
                      type: number
                  type: object
                type: array
          description: "Opération réussie"
        "400":
          description: "Erreur de paramètres"
        "500":
          description: Erreur technique
      summary: "Lister les réseaux de chaleur et les réseaux de froid"
      tags:
        - "Informations générales"
  "/v2/demands":
    get:
      description: "Renvoie toutes les demandes des réseaux rattachés à votre organisation, triées par date_modification croissante. Pour une synchro incrémentale, repassez la plus grande date_modification reçue en updated_since."
      parameters:
        - description: "Borne basse inclusive sur date_modification (ISO 8601). Omis : renvoie tout l'historique."
          in: query
          name: updated_since
          required: false
          schema:
            format: date-time
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  "$ref": "#/components/schemas/Demande"
                type: array
          description: Liste des demandes
        "401":
          description: "Token absent, invalide ou révoqué"
        "429":
          description: "Trop de requêtes"
      security:
        - bearerAuth: []
      summary: Lister les demandes de raccordement de votre organisation
      tags:
        - API partenaire
  "/v2/demands/{id}":
    patch:
      description: "Seuls statut et commentaire sont modifiables ; les autres champs sont gérés par FCU et ignorés."
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: uuid
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                commentaire:
                  description: Commentaire de suivi du gestionnaire.
                  anyOf:
                    - type: string
                    - type: "null"
                statut:
                  type: string
                  enum:
                    - "À traiter"
                    - "Non réalisable"
                    - "Recontacté pour étude"
                    - "Proposition commerciale envoyée"
                    - "Voté en AG"
                    - Travaux en cours
                    - "Réalisé"
                    - "Projet abandonné par le prospect"
                  description: "Statut d'avancement : À traiter, Non réalisable, Recontacté pour étude, Proposition commerciale envoyée, Voté en AG, Travaux en cours, Réalisé, Projet abandonné par le prospect."
              additionalProperties: false
              minProperties: 1
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                type: object
          description: "Mise à jour effectuée (la demande n'est pas renvoyée)"
        "401":
          description: "Token absent, invalide ou révoqué"
        "404":
          description: "Demande introuvable ou hors de votre périmètre"
        "429":
          description: "Trop de requêtes"
      security:
        - bearerAuth: []
      summary: "Mettre à jour le suivi d'une demande"
      tags:
        - API partenaire
servers:
  - url: "https://france-chaleur-urbaine.beta.gouv.fr/api"
tags:
  - name: "Informations générales"
  - name: API partenaire